% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_pca.R
\name{graph_pca}
\alias{graph_pca}
\title{Visualize PCA results}
\usage{
graph_pca(pca_result, color_data = NULL, labels = NULL)
}
\arguments{
\item{pca_result}{PCA result object.}

\item{color_data}{Optional vector specifying color for data points.}

\item{labels}{Optional vector of labels for data points.}
}
\value{
A ggplot object representing the PCA scatter plot.
}
\description{
This function visualizes the results of Principal Component Analysis (PCA) using a scatter plot.
}
