% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load RNA-seq and clinical data}
\usage{
load_data(RNAseq_path, Clinical_path)
}
\arguments{
\item{RNAseq_path}{Path to the CSV file containing RNA-seq data.}

\item{Clinical_path}{Path to the CSV file containing clinical data.}
}
\value{
None
}
\description{
This function reads RNA-seq and clinical data from CSV files and assigns them to the global environment.
}
