% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_and_zscore}
\alias{normalize_and_zscore}
\title{Normalize and compute Z-scores}
\usage{
normalize_and_zscore(data)
}
\arguments{
\item{data}{Input data matrix.}
}
\value{
A matrix of Z-scores.
}
\description{
This function performs log2 normalization, quantile normalization, and computes Z-scores for the input data.
}
