% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_kmeans.R
\name{perform_kmeans_with_silhouette}
\alias{perform_kmeans_with_silhouette}
\title{Perform K-means clustering with silhouette analysis}
\usage{
perform_kmeans_with_silhouette(pca_result, max_clusters)
}
\arguments{
\item{pca_result}{PCA result object.}

\item{max_clusters}{Maximum number of clusters to consider.}
}
\value{
K-means clustering result object.
}
\description{
This function performs K-means clustering with silhouette analysis to determine the optimal number of clusters.
}
