% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_pca.R
\name{perform_pca}
\alias{perform_pca}
\title{Perform Principal Component Analysis (PCA)}
\usage{
perform_pca(data)
}
\arguments{
\item{data}{Input data matrix.}
}
\value{
PCA result object.
}
\description{
This function performs Principal Component Analysis (PCA) on the input data matrix.
}
