% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kmeans.R
\name{plot_kmeans_results}
\alias{plot_kmeans_results}
\title{Visualize K-means clustering results}
\usage{
plot_kmeans_results(pca_result, kmeans_result)
}
\arguments{
\item{pca_result}{PCA result object.}

\item{kmeans_result}{K-means clustering result object.}
}
\value{
A ggplot object representing the K-means clustering results.
}
\description{
This function visualizes the results of K-means clustering using a scatter plot.
}
