% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_unshared_rows.R
\name{remove_unshared_rows}
\alias{remove_unshared_rows}
\title{Remove rows not shared between two data frames}
\usage{
remove_unshared_rows(df1, df2)
}
\arguments{
\item{df1}{First data frame.}

\item{df2}{Second data frame.}
}
\value{
Merged data frame with shared rows.
}
\description{
This function removes rows that are not shared between two data frames and returns a merged data frame.
}
